+++ Sources/OpenCoreAnimation/CAKeyframeAnimation.swift
@@ -19,7 +19,7 @@ open class CAKeyframeAnimation: CAPropertyAnimation {
 
     /// An optional array of numbers that define the time at which each keyframe value is applied.
     /// Values should be in the range [0, 1].
-    open var keyTimes: [Float]?
+    open var keyTimes: [CGFloat]?
 
     /// An optional array of CAMediaTimingFunction objects.
     open var timingFunctions: [CAMediaTimingFunction]?
@@ -29,15 +29,15 @@ open class CAKeyframeAnimation: CAPropertyAnimation {
 
     /// An array of numbers that define the tightness of the curve.
     /// Used with cubic and Catmull-Rom calculation modes.
-    open var tensionValues: [Float]?
+    open var tensionValues: [CGFloat]?
 
     /// An array of numbers that define the sharpness of timing curve corners.
     /// Used with cubic calculation mode.
-    open var continuityValues: [Float]?
+    open var continuityValues: [CGFloat]?
 
     /// An array of numbers that define the position of the curve relative to a control point.
     /// Used with cubic calculation mode.
-    open var biasValues: [Float]?
+    open var biasValues: [CGFloat]?
 
     /// Determines whether objects animating along the path rotate to match the path tangent.
     open var rotationMode: CAAnimationRotationMode?
